+++ llm_web_kit/config/pipe_tpl/html-test.jsonc
@@ -23,7 +23,7 @@
         "extractor": [
             {
                 "enable": true,
-                "python_class": "llm_web_kit.extractor.html.extractor.HTMLFileFormatExtractor",
+                "python_class": "llm_web_kit.extractor.html.extractor.MagicHTMLFIleFormatorExtractor",
                 "class_init_kwargs": {}
             }
         ],
