+++ llm_web_kit/config/pipe_tpl/noclip_html.jsonc
@@ -16,7 +16,7 @@
             "extractor": [
                 {
                     "enable": true,
-                    "python_class": "llm_web_kit.extractor.html.extractor.MagicHTMLFIleFormatorExtractor",
+                    "python_class": "llm_web_kit.extractor.html.extractor.NoClipHTMLFIleFormatorExtractor",
                     "class_init_kwargs": {}
                 }
             ],
