+++ llm_web_kit/simple.py
@@ -5,9 +5,9 @@ from datetime import datetime
 
 from llm_web_kit.config.cfg_reader import load_pipe_tpl
 from llm_web_kit.extractor.extractor_chain import ExtractSimpleFactory
-from llm_web_kit.extractor.html.extractor import HTMLFileFormatExtractor
-from llm_web_kit.extractor.html.pure_extractor import \
-    PureHTMLFileFormatExtractor
+from llm_web_kit.extractor.html.extractor import (
+    HTMLPageLayoutType, MagicHTMLFIleFormatorExtractor,
+    NoClipHTMLFIleFormatorExtractor)
 from llm_web_kit.input.datajson import DataJson
 
 
@@ -35,8 +35,8 @@ class ExtractorFactory:
             raise ValueError(f'Invalid extractor type: {extractor_type}')
 
 
-def __extract_pure_html(url:str, html_content: str) -> DataJson:
-    extractor = PureHTMLFileFormatExtractor(load_pipe_tpl('html'))
+def __extract_main_html_by_no_clip_html(url:str, html_content: str) -> DataJson:
+    extractor = NoClipHTMLFIleFormatorExtractor(load_pipe_tpl('noclip_html'))
     input_data_dict = {
         'track_id': str(uuid.uuid4()),
         'url': url,
@@ -52,8 +52,8 @@ def __extract_pure_html(url:str, html_content: str) -> DataJson:
     return result
 
 
-def __extract_magic_html(url:str, html_str: str, page_layout_type:str) -> DataJson:
-    magic_html_extractor = HTMLFileFormatExtractor(load_pipe_tpl('html'))
+def __extract_main_html_by_maigic_html(url:str, html_str: str, page_layout_type:str) -> DataJson:
+    magic_html_extractor = MagicHTMLFIleFormatorExtractor(load_pipe_tpl('html'))
     main_html, method, title = magic_html_extractor._extract_main_html(html_str, url, page_layout_type)
     return main_html, title
 
@@ -74,31 +74,25 @@ def __extract_html(url:str, html_content: str) -> DataJson:
     return result
 
 
-def extract_pure_html_to_md(url:str, html_content: str) -> str:
-    """extract pure html to markdown without images."""
-    result = __extract_pure_html(url, html_content)
-    return result.get_content_list().to_nlp_md()
-
-
-def extract_pure_html_to_mm_md(url:str, html_content: str) -> str:
-    """extract pure html to markdown with images."""
-    result = __extract_pure_html(url, html_content)
-    return result.get_content_list().to_mm_md()
-
-
-def extract_html_to_md(url:str, html_content: str) -> str:
+def extract_html_to_md(url:str, html_content: str, clip_html=True) -> str:
     """extract html to markdown without images."""
-    result = __extract_html(url, html_content)
+    if clip_html:
+        result = __extract_html(url, html_content)
+    else:
+        result = __extract_main_html_by_no_clip_html(url, html_content)
     return result.get_content_list().to_nlp_md()
 
 
-def extract_html_to_mm_md(url:str, html_content: str) -> str:
+def extract_html_to_mm_md(url:str, html_content: str, clip_html=True) -> str:
     """extract html to markdown with images."""
-    result = __extract_html(url, html_content)
+    if clip_html:
+        result = __extract_html(url, html_content)
+    else:
+        result = __extract_main_html_by_no_clip_html(url, html_content)
     return result.get_content_list().to_mm_md()
 
 
-def extract_magic_html(url:str, html_str: str, page_layout_type:str = 'article') -> str:
+def extract_main_html_by_maigic_html(url:str, html_str: str, page_layout_type:str = HTMLPageLayoutType.LAYOUT_ARTICLE) -> str:
     """extract main html."""
-    result = __extract_magic_html(url, html_str, page_layout_type)
+    result = __extract_main_html_by_maigic_html(url, html_str, page_layout_type)
     return result[0], result[1]
