+++ llm_web_kit/tools/cli.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import click
 from loguru import logger
 
-from llm_web_kit.extractor.html.extractor import HTMLFileFormatExtractor
+from llm_web_kit.extractor.html.extractor import MagicHTMLFIleFormatorExtractor
 from llm_web_kit.input.datajson import DataJson
 
 
@@ -55,7 +55,7 @@ def cli(input_path, output_path, debug_mode):
             else:
                 raise ValueError('Input JSON must contain either html or path field')
 
-        extractor = HTMLFileFormatExtractor({})
+        extractor = MagicHTMLFIleFormatorExtractor({})
         data_e = extractor.extract(DataJson(input_data))
         output_json = data_e.to_json()
         if output_path:
