+++ tests/llm_web_kit/exception/test_exception_data.py
@@ -255,7 +255,7 @@ class TestException(unittest.TestCase):
                 'extractor': [
                     {
                         'enable': True,
-                        'python_class': 'llm_web_kit.extractor.html.extractor.HTMLFileFormatExtractor',
+                        'python_class': 'llm_web_kit.extractor.html.extractor.MagicHTMLFIleFormatorExtractor',
                         'class_init_kwargs': {},
                     }
                 ],
@@ -323,7 +323,7 @@ class TestException(unittest.TestCase):
                 'extractor': [
                     {
                         'enable': True,
-                        'python_class': 'llm_web_kit.extractor.html.extractor.HTMLFileFormatExtractor',
+                        'python_class': 'llm_web_kit.extractor.html.extractor.MagicHTMLFIleFormatorExtractor',
                         'class_init_kwargs': {},
                     }
                 ],
