+++ tests/llm_web_kit/extractor/test_extractor_chain_normal.py
@@ -145,7 +145,7 @@ class TestExtractorChainNormal(unittest.TestCase):
                 'extractor': [
                     {
                         'enable': False,
-                        'python_class': 'llm_web_kit.extractor.html.extractor.HTMLFileFormatExtractor',
+                        'python_class': 'llm_web_kit.extractor.html.extractor.MagicHTMLFIleFormatorExtractor',
                         'class_init_kwargs': {},
                     }
                 ],
@@ -226,7 +226,7 @@ class TestExtractorChainNormal(unittest.TestCase):
                 'extractor': [
                     {
                         'enable': True,
-                        'python_class': 'llm_web_kit.extractor.html.extractor.HTMLFileFormatExtractor',
+                        'python_class': 'llm_web_kit.extractor.html.extractor.MagicHTMLFIleFormatorExtractor',
                         'class_init_kwargs': {},
                     }
                 ]
